eventconsume/requirements.txt
PyYAML==3.11
argparse==1.2.1
oauthlib==1.0.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.4.0
wsgiref==0.1.2
